llm-server/requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
tokenizers
accelerate
bitsandbytes
sentencepiece
python-multipart
python-dotenv
httpx